taraul/requirements.txt
transformers
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
pypdf
sentence-transformers